team8/backend/requirements.txt
Django==4.2.27
djangorestframework==3.14.0
djangorestframework-gis==1.0
django-environ==0.11.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
requests==2.31.0
Pillow==10.2.0
django-filter==23.5

